+++ components/WalletContextProvider.tsx
@@ -1,9 +1,9 @@
 
 import { FC, ReactNode } from "react";
-import * as web3 from '@solana/web3.js'
 import { ConnectionProvider, WalletProvider } from '@solana/wallet-adapter-react'
 import { WalletModalProvider } from "@solana/wallet-adapter-react-ui";
-import { PhantomWalletAdapter } from '@solana/wallet-adapter-phantom';
+import * as web3 from '@solana/web3.js'
+import { PhantomWalletAdapter } from '@solana/wallet-adapter-wallets';
 require('@solana/wallet-adapter-react-ui/styles.css');
 
 const WalletContextProvider: FC<{ children: ReactNode }> = ({ children }) => {
