+++ pages/index.tsx
@@ -3,7 +3,7 @@ import styles from '../styles/Home.module.css'
 import WalletContextProvider from '../components/WalletContextProvider'
 import { AppBar } from '../components/AppBar'
 import { BalanceDisplay } from '../components/BalanceDisplay'
-import { SendSolForm } from '../components/SendSolForm'
+import { PingButton } from '../components/PingButton'
 import Head from 'next/head'
 
 const Home: NextPage = (props) => {
@@ -20,8 +20,7 @@ const Home: NextPage = (props) => {
       <WalletContextProvider>
         <AppBar />
         <div className={styles.AppBody}>
-          <BalanceDisplay />
-          <SendSolForm />
+          <PingButton />
         </div>
       </WalletContextProvider >
     </div>
